Seperated ui tests and non ui tests to different files
